extended the documentation for button keywords
Update issue 140
Status: Review
I also think that the XPath example is not really necessary, but I left the xpath in there because I refer to it in the two examples. If you think it's better to remove the XPath explanation from the documentation, I'd be okay with that decision.